include: Reorganize and de-duplicate common BPF headers
Consolidate core BPF header files from 'scheds/include/scx' and
'scheds/include/lib/alloc' into the top-level 'scheds/include' directory.

Key changes:
 - Remove duplicate headers for 'bpf_arena_spin_lock.h' and 'bpf_atomic.h'.
 - Relocate 'bpf_arena_common.bpf.h', 'bpf_arena_spin_lock.h',
   'bpf_atomic.h', and 'bpf_experimental.h' to 'scheds/include/'.
 - Update #include paths across the library and SCX schedulers (lavd,
   p2dq, rusty, and wd40) to reflect the new centralized structure.

This cleanup simplifies the include hierarchy and prevents logic
divergence between the BPF allocator and various scheduler components.
