+++ lib/ravg.bpf.c
@@ -1,7 +1,7 @@
 #include <scx/common.bpf.h>
 
-#include <scx/bpf_arena_common.bpf.h>
-#include <scx/bpf_arena_spin_lock.h>
+#include <bpf_arena_common.bpf.h>
+#include <bpf_arena_spin_lock.h>
 
 #include <lib/ravg.h>
