+++ lib/selftests/selftest.bpf.c
@@ -3,7 +3,7 @@
  * Copyright (c) 2025 Meta Platforms, Inc. and affiliates.
  */
 #include <scx/common.bpf.h>
-#include <scx/bpf_arena_common.bpf.h>
+#include <bpf_arena_common.bpf.h>
 
 #include <lib/rbtree.h>
