+++ scheds/include/lib/btree.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include <scx/common.bpf.h>
-#include <scx/bpf_arena_common.bpf.h>
-#include <scx/bpf_arena_spin_lock.h>
+#include <bpf_arena_common.bpf.h>
+#include <bpf_arena_spin_lock.h>
 
 #define BT_MAXLVL_PRINT (10)
 #define BT_LEAFSZ 10
