+++ scheds/include/lib/sdt_task.h
@@ -8,8 +8,8 @@
 
 #ifdef __BPF__
 
-#include <scx/bpf_arena_common.bpf.h>
-#include <scx/bpf_arena_spin_lock.h>
+#include <bpf_arena_common.bpf.h>
+#include <bpf_arena_spin_lock.h>
 #include "sdt_task_defs.h"
 #endif
