+++ scheds/rust/scx_lavd/src/bpf/balance.bpf.c
@@ -5,7 +5,7 @@
  */
 
 #include <scx/common.bpf.h>
-#include <scx/bpf_arena_common.bpf.h>
+#include <bpf_arena_common.bpf.h>
 #include "intf.h"
 #include "lavd.bpf.h"
 #include "util.bpf.h"
