+++ scheds/rust/scx_wd40/src/bpf/common.bpf.c
@@ -7,8 +7,8 @@
 #include <scx/common.bpf.h>
 #include <lib/sdt_task.h>
 
-#include <scx/bpf_arena_common.bpf.h>
-#include <scx/bpf_arena_spin_lock.h>
+#include <bpf_arena_common.bpf.h>
+#include <bpf_arena_spin_lock.h>
 
 #include <lib/cpumask.h>
