+++ scheds/rust/scx_wd40/src/bpf/deadline.bpf.c
@@ -10,8 +10,8 @@
 
 #include <lib/cpumask.h>
 
-#include <scx/bpf_arena_common.bpf.h>
-#include <scx/bpf_arena_spin_lock.h>
+#include <bpf_arena_common.bpf.h>
+#include <bpf_arena_spin_lock.h>
 
 #include "intf.h"
 #include "types.h"
