+++ scheds/rust/scx_wd40/src/bpf/lb_domain.bpf.c
@@ -10,8 +10,8 @@
 #include <lib/sdt_task.h>
 #include <lib/topology.h>
 
-#include <scx/bpf_arena_common.bpf.h>
-#include <scx/bpf_arena_spin_lock.h>
+#include <bpf_arena_common.bpf.h>
+#include <bpf_arena_spin_lock.h>
 
 #include <lib/cpumask.h>
