+++ scheds/rust/scx_wd40/src/bpf/main.bpf.c
@@ -41,8 +41,8 @@
 
 #include <scx/common.bpf.h>
 
-#include <scx/bpf_arena_common.bpf.h>
-#include <scx/bpf_arena_spin_lock.h>
+#include <bpf_arena_common.bpf.h>
+#include <bpf_arena_spin_lock.h>
 
 #include <lib/arena.h>
 #include <lib/cpumask.h>
