+++ scheds/rust/scx_wd40/src/bpf/placement.bpf.c
@@ -9,8 +9,8 @@
 
 #include <lib/cpumask.h>
 
-#include <scx/bpf_arena_common.bpf.h>
-#include <scx/bpf_arena_spin_lock.h>
+#include <bpf_arena_common.bpf.h>
+#include <bpf_arena_spin_lock.h>
 
 #include "intf.h"
 #include "types.h"
